+++ jdaviz/core/viewer_creators/image/image.py
@@ -13,7 +13,7 @@ class ImageViewerCreator(BaseViewerCreator):
     def __init__(self, app, **kwargs):
         super().__init__(app, **kwargs)
         self.dataset.filters = ['is_catalog_or_image_not_spectrum', 'is_not_wcs_only']
-        self.viewer_label_default = 'Image'  # TODO: append suffix if taken
+        self.viewer_label_default = 'Image'
 
     @property
     def viewer_class(self):