+++ CHANGES.rst
@@ -73,6 +73,8 @@ Other Changes and Additions
 
 - Add rotation keyword argument to `set_viewport` and `get_viewport` methods to expand the Astro Image Display API (AIDA) [#3759]
 
+- Scatter viewer support in the deconfigged app instance. [#3778]
+
 4.4.2 (unreleased)
 ==================